==> R/03_preprocessing_data.Rmd <==
---
title: "Preprocessing Raw Data"
author: "Yuxiao Li"
date: "`r Sys.Date()`"
output: github_document
---

```{r, echo = FALSE}
start_time <- Sys.time()
```

# Outline

We want to organize the dataset, RaindataPerepoch.csv,
in a tidy fashion to using ggplot. 
Basically, we want to

* redefine the first column data...time as the date and time object
* change the raw dataset as tidy data, each row corresponds to a single observation.
* save the processed as data as RData file "rain-tiny.RData".
* we consider four time intevals: 30 seconds, 1 minute, 15 minutes, and 1 hour.
* the raw (30-second) data are finally used for further application.

# Libraries

```{r load_libs, message=FALSE}

library(here)
library(reshape2)
library(ggplot2)
library(dplyr)
#library(cowplot)
```

# Change the data as dataframe
```{r process, warning=FALSE}

rain.raw <- read.csv(here("data","RaindataPerhour.csv"))
rain.long <- melt(rain.raw[,1:9], id=c("date...time"))
rain.long <- data.frame(rain.long)

names(rain.long)<-c("dt","sites","rain")
rain.long$dt <- as.POSIXct(strptime(rain.long$dt,format="%d.%m.%Y  %H:%M:%S"))
rain.long$Date <- as.Date(rain.long$dt)
rain.long$Time <- format((rain.long$dt) ,format = "%H:%M:%S")
```

# Plot the preprocessed data
```{r ggplot}
rainraw <- ggplot(rain.long,aes(y=rain,x=dt,group=sites,colour=sites))+
    geom_line()+facet_grid(sites ~ .) + xlab("Data and Time") + ylab("Rain rate (mm/hr)")
ggsave(filename=here("results","rainraw.pdf"), plot=rainraw, width = 11, height = 9)

```

```{r qqplot}
rainraw.pos<-rain.long[rain.long$rain>0,]
qq_raw.pos <- ggplot(rainraw.pos, aes(sample=rain)) + 
    stat_qq(alpha=0.5) + stat_qq_line() +
    xlab("Theoretical(mm/hr)") + ylab("Sample(mm/hr)")
ggsave(filename=here("results","qq_raw.pdf"), plot=qq_raw.pos, width = 7, height = 5)
```

```{r distributional table}
data_dist <- cut(unlist(rain.raw[-28801,2:9]),c(-0.1,0.1,2,5,10,33.6))
data_table = table(data_dist) ##The last point was removed to get 10-day data
names(data_table)<-c(" = 0", "(0,2]", "(2,5]", "(5,10]", "(10,33.6]")
data_table/28800/8
```



##Minutely Data-not run
```{r 1 Minute Data, eval=FALSE}
minutemean <- rain.long[,2:5] %>% group_by(G=trunc(2:(n()+1)/2))
rain.1min <- tapply(minutemean$rain,minutemean$G,sum)
rain.1min.df <- data.frame(dt = rain.long$dt[seq(1,dim(rain.long)[1],by=2)],sites = rain.long$sites[seq(1,dim(rain.long)[1],by=2)],rain = rain.1min)

ggplot(rain.1min.df,aes(y=rain,x=dt,group=sites,colour=sites))+
    geom_line()+facet_grid(sites ~ .) + xlab("Data and Time") + ylab("Rainfall Amount")
```

##15-Minute Data-not run
```{r 15 Minute Data, eval=FALSE}
minute15mean <- rain.long[,2:5] %>% group_by(G=trunc(30:(n()+29)/30))
rain.15min <- tapply(minute15mean$rain,minute15mean$G,sum)
rain.15min.df <- data.frame(dt = rain.long$dt[seq(1,dim(rain.long)[1],by=30)],sites = rain.long$sites[seq(1,dim(rain.long)[1],by=30)],rain = rain.15min)

ggplot(rain.15min.df,aes(y=rain,x=dt,group=sites,colour=sites))+
    geom_line()+facet_grid(sites ~ .) + xlab("Data and Time") + ylab("Rainfall Amount")
```


##Hourly Data-not run
```{r Hourly Data, eval=FALSE}
hourmean <- rain.long[,2:5] %>% group_by(G=trunc(120:(n()+119)/120))
rain.hour <- tapply(hourmean$rain,hourmean$G,sum)
rain.hour.df <- data.frame(dt = rain.long$dt[seq(1,dim(rain.long)[1],by=120)],sites = rain.long$sites[seq(1,dim(rain.long)[1],by=120)],rain = rain.hour)

ggplot(rain.hour.df,aes(y=rain,x=dt,group=sites,colour=sites))+
    geom_line()+facet_grid(sites ~ .) + xlab("Data and Time") + ylab("Rainfall Amount")
```

# Save the clean data

```{r save the file}
save(rain.long, file=here("results", "rain.long.RData"))
#save(rain.1min.df, file=here("results", "rain.1min.df.RData"))
#save(rain.15min.df, file=here("results", "rain.15min.df.RData"))
#save(rain.hour.df, file=here("results", "rain.hour.df.RData"))
```

## Running Time

```{r, echo = FALSE}
end_time <- Sys.time()

runtime_03 <- end_time-start_time

runtime_03
```
