==> R/11_simulation_study.Rmd <==
---
title: "Simulation Study"
author: "Yuxiao Li"
output: 
  github_document:
    toc: true
  html_document:
    toc: true
    number_sections: true
    code_folding: hide
---

```{r, echo = FALSE}
start_time <- Sys.time()
```

# Overview
This file contains the code for simulation studies.

In general, we want to

* examine the performance our model
* compare with a similar model with Gaussian error
* calculate the some important quantities related to rainfall
* visualize the results


## Libraries


```{r load_libs,message=FALSE}
library(here)
library(sn)
library(geoR)
library(ggplot2)
library(cowplot)
```

##Simulate synthetic dataset
### Function to simulate data

```{r simulate rainfall}
sims <- function(NT,NS,sigma21,beta1,b0,b1,alpha,nu){
   error.sim <- matrix(0,NT,NS)
   dist <- matrix(data = 0, nrow = NS, ncol = NS) # The distances between locations
    for (i in 1:NS) {
        for (j in 1:NS) {
            dist[i,j] <- sqrt((coord[i,2]-coord[j,2])^2+(coord[i,3]-coord[j,3])^2)   
        }
    }
    B1 <-  sigma21 * exp(-  dist/beta1)
    set.seed(1)
    for(s in 1 :NS){
        nu <- nu
        bvdelta <- sqrt(nu) * gamma(1/2 * (nu - 1))/sqrt(pi)/gamma(1/2 * nu) * 
            alpha/sqrt(1 + alpha^2)
        omega <- 1/sqrt((nu/(nu-2) - bvdelta^2))
        xi<- -omega * bvdelta
        error.sim[,s] <- rst(NT, xi = xi, omega = omega, alpha=alpha, nu=nu)
    }

    y.sim <- matrix(0,NT,NS)
    ut.sim <- ut[seq(1,28800,28800/NT),1:NS]
    for(t in 2: NT){
        y.sim[t,] <- B1 %*% y.sim[t - 1,] + error.sim[t,] * (b0 + b1 * mean(y.sim[t - 1,]))
        y.sim[t,] <- ifelse(y.sim[t,] > ut.sim[t,], y.sim[t,], 0)
    } 
    return(list(y.sim = y.sim, error.sim = error.sim))
}
```


### Parameter setups

```{r parameter setups}
load(here("results", "estimation.cov.VAR1.RData"))
NT <- 10000
NS <- 3
sigma21 <- 1/(NS+1)
beta1 <- 1
dist <- matrix(data = 0, nrow = NS, ncol = NS) # The distances between locations
for (i in 1:NS) {
    for (j in 1:NS) {
        dist[i,j] <- sqrt((coord[i,2]-coord[j,2])^2+(coord[i,3]-coord[j,3])^2)   
    }
}
B1 <-  sigma21 * exp(- dist/beta1)
b0 <- 0.5
b1 <- 0.5
alpha <- 3
nu <- 3
ut <- results.VAR1$ut
ut.sim <- ut[seq(1,28800,28800/NT),1:NS]
```

### Simulate one replicate
```{r simulate 1 error, eval = FALSE}
y.sim <- sims(NT,NS,sigma21,beta1,b0,b1,alpha,nu)
plot.ts(y.sim$y.sim)
```

## Visualize the synthetic data
### Simulate different cases wrt alpha and nu
```{r different alpha and nu}
y.sim.vis <- matrix(0,NT,6)
y.sim.vis[,1] <- sims(NT,NS,sigma21,beta1,b0,b1,0,20)$y.sim[,1]
y.sim.vis[,2] <- sims(NT,NS,sigma21,beta1,b0,b1,5,20)$y.sim[,1]
y.sim.vis[,3] <- sims(NT,NS,sigma21,beta1,b0,b1,0,7)$y.sim[,1]
y.sim.vis[,4] <- sims(NT,NS,sigma21,beta1,b0,b1,5,7)$y.sim[,1]
y.sim.vis[,5] <- sims(NT,NS,sigma21,beta1,b0,b1,0,3)$y.sim[,1]
y.sim.vis[,6] <- sims(NT,NS,sigma21,beta1,b0,b1,5,3)$y.sim[,1]



```

```{r visualization of alpha and nu}
pdf("../results/vis_nu_alpha.pdf",width=8, height=6)
par(mfrow=c(3,2),mar=c(1.5,4,1,1),mgp=c(1.5,0.5,0))
plot.ts(y.sim.vis[,1], ylim=c(0,max(y.sim.vis)),
        ylab = expression(paste(nu,"=20")),
        main = expression(paste(alpha,"=0")))
plot.ts(y.sim.vis[,2], ylim=c(0,max(y.sim.vis)),
        ylab = '', main=expression(paste(alpha,"=5")))
plot.ts(y.sim.vis[,3], ylim=c(0,max(y.sim.vis)), 
        ylab = expression(paste(nu,"=7")))
plot.ts(y.sim.vis[,4], ylim=c(0,max(y.sim.vis)), 
        ylab='')
plot.ts(y.sim.vis[,5], ylim=c(0,max(y.sim.vis)),
        ylab=expression(paste(nu,"=3")))
plot.ts(y.sim.vis[,6], ylim=c(0,max(y.sim.vis)),
        ylab='')
dev.off()
```


### Simulate different cases wrt b0 and b1

```{r different b0 and b1,eval = FALSE}
y.sim.vis2 <- matrix(0,NT,6)
y.sim.vis2[,1] <- sims(NT,NS,sigma21,beta1,0.5,0.2,0,20)$y.sim[,1]
y.sim.vis2[,2] <- sims(NT,NS,sigma21,beta1,0.7,0.2,0,20)$y.sim[,1]
y.sim.vis2[,3] <- sims(NT,NS,sigma21,beta1,0.5,0.5,0,20)$y.sim[,1]
y.sim.vis2[,4] <- sims(NT,NS,sigma21,beta1,0.7,0.5,0,20)$y.sim[,1]
y.sim.vis2[,5] <- sims(NT,NS,sigma21,beta1,0.5,0.8,0,20)$y.sim[,1]
y.sim.vis2[,6] <- sims(NT,NS,sigma21,beta1,0.7,0.8,0,20)$y.sim[,1]
```

### Simulate different cases
```{r visualization of b0 and b1, eval = FALSE}
pdf("../results/vis_b0_b1.pdf",width=8, height=6)
par(mfrow=c(3,2),mar=c(1.5,4,1,1),mgp=c(1.5,0.5,0))
plot.ts(y.sim.vis2[,1], ylim=c(0,max(y.sim.vis)),
        ylab = expression(paste(b[1],"=0.2")),
        main = expression(paste(b[0],"=0.5")))
plot.ts(y.sim.vis2[,2], ylim=c(0,max(y.sim.vis2)),
        ylab = '', main=expression(paste(b[0],"=0.7")))
plot.ts(y.sim.vis2[,3], ylim=c(0,max(y.sim.vis2)), 
        ylab = expression(paste(b[1],"=0.5")))
plot.ts(y.sim.vis2[,4], ylim=c(0,max(y.sim.vis2)), 
        ylab='')
plot.ts(y.sim.vis2[,5], ylim=c(0,max(y.sim.vis2)),
        ylab=expression(paste(b[1],"=0.8")))
plot.ts(y.sim.vis2[,6], ylim=c(0,max(y.sim.vis2)),
        ylab='')
dev.off()
```


## Estimate all other parameters - with spatial covariance var 1-st
### likelihood function
```{r lik fun}
est.sims<-function(y.sim, nu){
    NT <- nrow(raindata)
    NS <- ncol(raindata)

    init <- c()
    #init[1] <- 0.1/(1/NS-0.1) #variance
    init[1] <- 0.1
    init[2] <- 1 #spatial range
    init[3] <- .5
    init[4] <- .5
    init[5] <- 1 #positive skewness alpha
    init <- log(init)
    cov.optim.rslt.st <- optim(par=init,fn = varlik.sp.var1.st,method="BFGS",
                    data = y.sim, coord = coord, ut = ut.sim, nu = nu)
    st.sigma21 <- exp(cov.optim.rslt.st$par[1])/(1+exp(cov.optim.rslt.st$par[1]))
    st.beta1 <- exp(cov.optim.rslt.st$par[2])
    st.b0 <- exp(cov.optim.rslt.st$par[3])
    st.b1 <- exp(cov.optim.rslt.st$par[4])
    st.alpha <- exp(cov.optim.rslt.st$par[5])
    est_st <- c(st.sigma21,st.beta1,st.b0,st.b1,st.alpha)
    
    init_norm <- init[-5]
    cov.optim.rslt.norm <- optim(par=init_norm,fn = varlik.sp.var1.norm,
                                 method = "BFGS",data = y.sim, coord = coord, 
                                 ut = ut.sim)
    norm.sigma21 <- exp(cov.optim.rslt.norm$par[1])/
        (1+exp(cov.optim.rslt.norm$par[1]))
    norm.beta1 <- exp(cov.optim.rslt.norm$par[2])
    norm.b0 <- exp(cov.optim.rslt.norm$par[3])
    norm.b1 <- exp(cov.optim.rslt.norm$par[4])
    est_n<-c(norm.sigma21,norm.beta1,norm.b0,norm.b1)
    return(list(est_st=est_st,est_n=est_n))
}

```


### Run the estimations
```{r fit}
y.sim <- sims(10000,3,sigma21,beta1,b0,b1,0,3)$y.sim
est1 <- est.sims(y.sim=y.sim,nu=3)
y.sim2 <- sims(10000,3,sigma21,beta1,b0,b1,5,3)$y.sim
est2 <- est.sims(y.sim=y.sim2,nu=3)
y.sim3 <- sims(10000,3,sigma21,beta1,b0,b1,0,7)$y.sim
est3 <- est.sims(y.sim=y.sim3,nu=7)
y.sim4 <- sims(10000,3,sigma21,beta1,b0,b1,5,7)$y.sim
est4 <- est.sims(y.sim=y.sim4,nu=7)
y.sim5 <- sims(10000,3,sigma21,beta1,b0,b1,0,20)$y.sim
est5 <- est.sims(y.sim=y.sim5,nu=20)
y.sim6 <- sims(10000,3,sigma21,beta1,b0,b1,5,20)$y.sim
est6 <- est.sims(y.sim=y.sim6,nu=20)
```

```{r}
for (i in 1:10){
y.sim <- sims(10000,3,sigma21,beta1,b0,b1,0,3)$y.sim
est1 <- est.sims(y.sim=y.sim,nu=3)
y.sim2 <- sims(10000,3,sigma21,beta1,b0,b1,5,3)$y.sim
est2 <- est.sims(y.sim=y.sim2,nu=3)
y.sim3 <- sims(10000,3,sigma21,beta1,b0,b1,0,7)$y.sim
est3 <- est.sims(y.sim=y.sim3,nu=7)
y.sim4 <- sims(10000,3,sigma21,beta1,b0,b1,5,7)$y.sim
est4 <- est.sims(y.sim=y.sim4,nu=7)
y.sim5 <- sims(10000,3,sigma21,beta1,b0,b1,0,20)$y.sim
est5 <- est.sims(y.sim=y.sim5,nu=20)
y.sim6 <- sims(10000,3,sigma21,beta1,b0,b1,5,20)$y.sim
est6 <- est.sims(y.sim=y.sim6,nu=20)
mat.est<-cbind(unlist(est1),unlist(est2),unlist(est3),unlist(est4),unlist(est5),unlist(est6))
save(mat.est, file = here("results",paste0("sim.est",i,".RData")))
}

```

### Save the estimation
```{r save est}
mat.est<-cbind(unlist(est1),unlist(est2),unlist(est3),unlist(est4),unlist(est5),unlist(est6))
save(mat.est, file = here("results","sim.est.RData"))
```

## 50 bootstrap samples to quantify the uncertainties
### Functions to generate samples
```{r multiple simulations}

nsims <- function(NT,NS,sigma21,beta1,b0,b1,alpha,nu,niters){
    error <- array(0,c(NT,NS,niters))
    y <- array(0,c(NT,NS,niters))
    dist <- matrix(data = 0, nrow = NS, ncol = NS) # The distances between locations
    for (i in 1:NS) {
        for (j in 1:NS) {
            dist[i,j] <- sqrt((coord[i,2]-coord[j,2])^2+(coord[i,3]-coord[j,3])^2)   
        }
    }
    B1 <-  sigma21 * exp(-dist/beta1)
    for(iter in 1: niters){
        for(s in 1 :NS){
            nu <- nu
            bvdelta <- sqrt(nu) * gamma(1/2 * (nu - 1))/sqrt(pi)/gamma(1/2 * nu) * 
            alpha/sqrt(1 + alpha^2)
            omega <- 1/sqrt((nu/(nu-2) - bvdelta^2))
            xi<- -omega * bvdelta
            error[,s,iter] <- rst(NT, xi = xi, omega = omega, alpha, nu)
        }

        for(t in 2: NT){
        y[t,,iter] <- B1 %*% y[t - 1,,iter] + 
                            error[t,,iter] * (b0 + b1 * mean(y[t - 1,,iter]))
        y[t,,iter] <- ifelse(y[t,,iter] > ut.sim[t,], y[t,,iter], 0)
        }
    }
    return(y)
}
```


### Generate samples

```{r reproduce}
load(here("results","sim.est.RData"))
nu.iter<-c(3,3,7,7,20,20)
alpha.iter<-c(0,5,0,5,0,5)
niters<-50
rmse.st<-rmse.norm<-matrix(0,6,niters)
true.sim<-array(0,c(NT,NS,6))
est.sim<-est.sim.norm<-array(0,c(NT,NS,niters,6))
for(i in 1:6){
    true.sim[,,i] <- sims(10000, 3, sigma21, beta1, b0, b1,
                          alpha.iter[i], nu.iter[i])$y.sim
    est.sim[,,,i] <- nsims(10000, 3, mat.est[1,i],
                           mat.est[2,i], mat.est[3,i],
                           mat.est[4,i], mat.est[5,i],
                           nu.iter[i], niters)
    est.sim.norm[,,,i] <- nsims(10000, 3, mat.est[6,i],
                                mat.est[7,i], mat.est[8,i],
                                mat.est[9,i], 0, 150, niters)
}

```

### Calculate MRMSEs 
```{r mrmse}
for(i in 1:6){
    for(j in 1:niters){
        rmse.st[i,j] <- sqrt(mean((est.sim[,,j,i]-true.sim[,,i])^2))
        rmse.norm[i,j] <- sqrt(mean((est.sim.norm[,,j,i]-true.sim[,,i])^2))
    }
}
mrmse.st<-apply(rmse.st,1,mean)*100
mrmse.norm<-apply(rmse.norm,1,mean)*100
mrmse.norm/mrmse.st
```

### Simple QQplot
```{r qqplot,eval = FALSE}
par(mfrow=c(2,3))
for(i in 1:6){
qqplot(est.sim[,1,,i],true.sim[,1,i])
abline(0,1)
qqplot(est.sim.norm[,1,,i],true.sim[,1,i])
abline(0,1) 
}
```

### QQplot with uncertainties
```{r calculate quantiles}
quan<-quan.norm<-matrix(0,niters,1000)
for (j in 1 :niters){
   quan[j,]<- quantile(est.sim[,1,j,2],1:1000/1000) 
   quan.norm[j,]<- quantile(est.sim.norm[,1,j,2],1:1000/1000) 
   quan.true<- quantile(true.sim[,1,2],1:1000/1000) 
}
quan.df<-data.frame(x=rep(quan.true,3),median=c(apply(quan,2,sort)[26,],apply(quan.norm,2,sort)[26,],quan.true),lower=c(apply(quan,2,sort)[2,],apply(quan.norm,2,sort)[2,],rep(NA,1000)),upper=c(apply(quan,2,sort)[49,],apply(quan.norm,2,sort)[49,],rep(NA,1000)),Type=c(rep("Skew-t",1000),rep("Gaussian",1000),rep("Truth",1000)))

for (j in 1 :niters){
   quan[j,]<- quantile(est.sim[,1,j,5],1:1000/1000) 
   quan.norm[j,]<- quantile(est.sim.norm[,1,j,5],1:1000/1000) 
   quan.true<- quantile(true.sim[,1,5],1:1000/1000) 
}
quan.df2<-data.frame(x=rep(quan.true,3),median=c(apply(quan,2,sort)[26,],apply(quan.norm,2,sort)[26,],quan.true),lower=c(apply(quan,2,sort)[2,],apply(quan.norm,2,sort)[2,],rep(NA,1000)),upper=c(apply(quan,2,sort)[49,],apply(quan.norm,2,sort)[49,],rep(NA,1000)),Type=c(rep("Skew-t",1000),rep("Gaussian",1000),rep("Truth",1000)))
```

### Draw QQplots 
```{r plot qq 3 and 5}
p <- ggplot(data = quan.df, aes(x=x, y=median, colour=Type)) + 
    geom_line(size=0.8) + ylab("Quantile") + xlab("Quantile") +
    ggtitle(expression(paste("(a) QQ-plot of the two models when ",
                             nu,"=3, ",alpha,"=5"))) +
    scale_y_continuous(limits = c(1,7))+scale_x_continuous(limits = c(1,5))+
    geom_ribbon(aes(ymin=lower, ymax=upper,fill=Type), linetype=2, alpha=0.3) + 
    scale_color_brewer(palette = "Dark2") +    
    scale_fill_brewer(palette = "Dark2")
p
```

```{r plot qq 20 and 0}
p2 <- ggplot(data = quan.df2, aes(x=x, y=median, colour=Type)) + 
    geom_line(size=0.8) + ylab("Quantile") + xlab("Quantile") +
    ggtitle(expression(paste("(b) QQ-plot of the two models when ",
                             nu,"=20, ",alpha,"=0"))) +
    scale_y_continuous(limits = c(1,6))+scale_x_continuous(limits = c(1,4.2))+
    geom_ribbon(aes(ymin=lower, ymax=upper,fill=Type), linetype=2, alpha=0.3) + 
    scale_color_brewer(palette = "Dark2") +    
    scale_fill_brewer(palette = "Dark2")
p2
```

```{r save the figure}
combine_plot<-plot_grid(p, p2, align='h',rel_widths = 1.2, rel_heights = 1)
ggsave(filename=here("results","qqsim.pdf"),device='pdf', plot=combine_plot, width = 12, height = 5)
```

## Calculate the dry probability
###Calculate
```{r dry prob}
dp.true<-sum(true.sim[,,2]==0)/length(true.sim[,,2])
dp.st<-dp.norm<-c()
for(i in 1: niters) {
    dp.st[i]<-sum(est.sim[,,i,2]==0)/length(est.sim[,,i,2])
    dp.norm[i]<-sum(est.sim.norm[,,i,2]==0)/length(est.sim.norm[,,i,2])
}

```

Plot the results
```{r draw and save}
pdf("../results/dryprob.pdf",width=8, height=6)
plot(density(dp.st),xlim=c(0.9,1),xlab="Probability",col=2,lwd=2,lty=2,main="Dry Probability")
abline(v=dp.true,col=1,lwd=2,lty=1)
lines(density(dp.norm),col=4,lwd=2,lty=3)
legend("topleft",c("Truth","Skew-t Error","Gaussian Error"),col=c(1,2,4),lwd=2,lty=c(1,2,3))
dev.off()
```

## Running time
```{r, echo = FALSE}
end_time <- Sys.time()

runtime_11 <- end_time-start_time

runtime_11
```
